Moving Import private functions to their own module to make testing and debugging the internal steps easier
